pyproject.toml
[project]
name = "sgrades"
version = "0.1.0"
description = "S-GRADES - Student Grading and Assessment System"
requires-python = ">3.13,<3.14"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn==0.32.0",
    "starlette==0.38.6",
    "pydantic==2.10.0",
    "python-dotenv==1.0.1",
    "SQLAlchemy==2.0.41",
    "pandas==2.3.1",
    "numpy==2.1.3",
    "scikit-learn==1.7.1",
    "scipy==1.15.3",
    "datasets==2.19.1",
    "huggingface-hub==0.34.4",
    "requests==2.32.3",
    "httpx==0.28.1",
    "slowapi==0.1.9",
    "python-multipart==0.0.20",
    "PyYAML==6.0.2",
    "tqdm==4.67.1",
    "psycopg2 (>=2.9.10,<3.0.0)",
]

[project.optional-dependencies]
dev = [
    "ipython",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"